fe-feat: added initial meta-data transformation. And refactored a bit.
