+++ packages/core/transformations/transformation.typedef.js
@@ -1,0 +1,11 @@
+/**
+ * @callback TypeFnTransform
+ * @param {string} html
+ * @return {Promise<string>}
+ */
+
+/**
+ * @typedef Transformation
+ * @property {TypeFnTransform} transform
+ */
+
