+++ packages/core/typedef.js
@@ -1,0 +1,8 @@
+/**
+ * @typedef {import('./define-config.js').McFlyConfig} Config
+ * @typedef {import('unstorage').Storage} Storage
+ * @typedef {import('unstorage').StorageValue} StorageValue
+ * @typedef {import('ultrahtml').Node} HtmlNode
+ * @typedef {import('estree').BaseNode} JsNode
+ * @typedef {import('h3').EventHandler} EventHandler
+ */
